Changes to pywcs tests:
test_maps() and test_spectra() use glob to find the files to use for the tests.

- Change the file name handling so only the base name of the file appears in the test name, instead of the full path name.  (Every time something changes about a directory name, all the test names change and I see hundreds of missing tests.)

- Check that the correct number of test files are found.  If *.hdr does not match anything, the old implementation silently did nothing.  I only noticed because I was specifically looking for the tests that did not show up.

test_all_units() makes non-unique test names.  Pandokia test names are not case sensitive, so a test named "d" and "D" are the same test.  To fix that, I added "-1", "-2", etc to ambiguous test names.



git-svn-id: https://aeon.stsci.edu/ssb/svn/astrolib/trunk/pywcs@2234 90a0a646-be8a-0410-bb88-9290da87bc01

Former-commit-id: 4639e7a3419efa50125ded42a8652978e4c2431e